CC/Linker: Move to own TU, and remove dependency
